+++ packages/liferay-npm-scripts/support/jest/matchers.js
@@ -1,0 +1,7 @@
+/**
+ * © 2019 Liferay, Inc. <https://liferay.com>
+ *
+ * SPDX-License-Identifier: BSD-3-Clause
+ */
+
+require('./matchers/toLintStyles');
